Move random_group_level para tools.py
